Wrote quick heapsort implementation.
The implementation of heapsort needs a bit of refinement, but it works (yay!). TODO on this program still includes allowing switching from maxheap to minheap, fixing some naming issues, and fixing the templates to be a lot more user friendly. Needing to enter the type of the sorter (which is swapped from the final result) and then create a new instance of it as an argument is clunky.